[TRUNK][FEATURE] - adds export capabilities for task comments - minor changes in conversion classes
git-svn-id: https://svn.libreccm.org/ccm/trunk@5013 8810af33-2d31-482b-a856-94f89814c4df

+++ ccm-core/src/com/arsdigita/portation/conversion/core/security/PermissionConversion.java
@@ -81,7 +81,7 @@ public class PermissionConversion {
      */
     private static void createPermissionsAndSetAssociations(final List<com
             .arsdigita.kernel.permissions.Permission> trunkPermissions) {
-        long processed = 0, skipped = 0;
+        int processed = 0, skipped = 0;
 
         for (com.arsdigita.kernel.permissions.Permission trunkPermission :
                 trunkPermissions) {
@@ -150,7 +150,7 @@ public class PermissionConversion {
      */
     private static void setGranteeDependency(final List<com.arsdigita.kernel
             .permissions.Permission> trunkPermissions) {
-        long duplicates = 0;
+        int duplicates = 0;
 
         for (com.arsdigita.kernel.permissions.Permission trunkPermission :
                 trunkPermissions) {
